#169 refactor: CustomUserDetailsService의 validateIfNewUser 메서드를 UserService로 이동 및 탈퇴 여부 판별 로직 추가
