+++ src/main/java/com/floney/floney/user/controller/KakaoController.java
@@ -1,7 +1,6 @@
 package com.floney.floney.user.controller;
 
 import com.floney.floney.user.dto.request.SignupRequest;
-import com.floney.floney.user.service.CustomUserDetailsService;
 import com.floney.floney.user.service.KakaoUserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -17,7 +16,6 @@ import org.springframework.web.bind.annotation.RestController;
 public class KakaoController implements AuthController {
 
     private final KakaoUserService kakaoUserService;
-    private final CustomUserDetailsService customUserDetailsService;
 
     @Override
     @GetMapping("/check")
@@ -28,7 +26,6 @@ public class KakaoController implements AuthController {
     @Override
     @PostMapping("/signup")
     public ResponseEntity<?> signup(String token, SignupRequest request) {
-        customUserDetailsService.validateIfNewUser(request.getEmail());
         kakaoUserService.signup(token, request);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
